Add Config Support
In this commit i have added config.py file which handles all the configuration of the application , It takes all the configuration from .env file.
In order to achive this i have installed pydatic settings libray .
From fruther onwards if we require config anywhere in application we just need to call config function.